proces_hp_package N°22/requirements.txt
pandas
matplotlib
python-docx
tqdm
regex
nltk
spacy